Add weights into wave function collapse. UI improvements.

+++ MarsUndiscovered.Game/Components/WaveFunction/TileAttribute.cs
@@ -3,7 +3,7 @@
 public class TileAttribute
 {
     public string Symmetry { get; set; }
-    public double Weight { get; set; }
+    public int Weight { get; set; }
     public string Adapters { get; set; }
     public bool EnableMirror { get; set; }
 }
